@@ -28,61 +28,60 @@
     "version": "conventional-changelog -i CHANGELOG.md -s -r 0 && git add CHANGELOG.md"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.8",
-    "@angular/animations": "~9.1.11",
-    "@angular/cli": "~9.1.8",
-    "@angular/common": "~9.1.11",
-    "@angular/compiler": "~9.1.11",
-    "@angular/compiler-cli": "~9.1.11",
-    "@angular/core": "~9.1.11",
-    "@angular/forms": "~9.1.11",
-    "@angular/platform-browser": "~9.1.11",
-    "@angular/platform-browser-dynamic": "~9.1.11",
-    "@angular/router": "~9.1.11",
-    "@fortawesome/angular-fontawesome": "^0.6.1",
-    "@fortawesome/fontawesome-svg-core": "^1.2.28",
-    "@fortawesome/free-solid-svg-icons": "^5.13.0",
-    "@ngx-loading-bar/core": "^4.2.0",
-    "@ngx-loading-bar/http-client": "^4.2.0",
-    "@ngx-loading-bar/router": "^4.2.0",
-    "@ngx-translate/core": "^12.1.2",
-    "@ngx-translate/http-loader": "^4.0.0",
-    "@types/jasmine": "~3.5.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "bootstrap": "^4.5.0",
-    "codelyzer": "^5.1.2",
-    "electron": "9.0.4",
-    "electron-builder": "22.7.0",
+    "@angular-devkit/build-angular": "^14.2.8",
+    "@angular/animations": "~14.2.9",
+    "@angular/cli": "~14.2.8",
+    "@angular/common": "~14.2.9",
+    "@angular/compiler": "~14.2.9",
+    "@angular/compiler-cli": "~14.2.9",
+    "@angular/core": "~14.2.9",
+    "@angular/forms": "~14.2.9",
+    "@angular/platform-browser": "~14.2.9",
+    "@angular/platform-browser-dynamic": "~14.2.9",
+    "@angular/router": "~14.2.9",
+    "@fortawesome/angular-fontawesome": "^0.11.1",
+    "@fortawesome/fontawesome-svg-core": "^6.2.0",
+    "@fortawesome/free-solid-svg-icons": "^6.2.0",
+    "@ngx-loading-bar/core": "^6.0.2",
+    "@ngx-loading-bar/http-client": "^6.0.2",
+    "@ngx-loading-bar/router": "^6.0.2",
+    "@ngx-translate/core": "^14.0.0",
+    "@ngx-translate/http-loader": "^7.0.0",
+    "@types/jasmine": "~4.3.0",
+    "@types/jasminewd2": "~2.0.10",
+    "@types/node": "^18.11.9",
+    "bootstrap": "^5.2.2",
+    "codelyzer": "^6.0.2",
+    "electron": "21.2.2",
+    "electron-builder": "23.6.0",
     "electron-reload": "1.5.0",
-    "handlebars": "^4.7.6",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~5.0.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~3.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "ngx-infinite-scroll": "^9.0.0",
+    "handlebars": "^4.7.7",
+    "karma": "~6.4.1",
+    "karma-chrome-launcher": "~3.1.1",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "ngx-infinite-scroll": "^14.0.0",
     "npm-run-all": "^4.1.5",
     "protractor": "~7.0.0",
-    "rxjs": "~6.5.4",
-    "ts-node": "~8.3.0",
-    "tslib": "^1.10.0",
+    "rxjs": "~7.5.7",
+    "ts-node": "~10.9.1",
+    "tslib": "^2.4.1",
     "tslint": "~6.1.0",
-    "typescript": "~3.8.3",
-    "wait-on": "^5.0.1",
-    "webtorrent": "0.108.6",
-    "zone.js": "~0.10.2"
+    "typescript": "~4.8.4",
+    "wait-on": "^6.0.1",
+    "webtorrent": "1.9.1",
+    "zone.js": "~0.11.8"
   },
   "dependencies": {
-    "custom-electron-titlebar": "3.2.2-hotfix62",
-    "electron-log": "4.2.2",
-    "electron-updater": "4.3.1",
-    "inversify": "5.0.1",
+    "@electron/remote": "^2.0.8",
+    "custom-electron-titlebar": "^4.1.0",
+    "electron-log": "4.4.8",
+    "electron-updater": "5.3.0",
+    "inversify": "6.0.1",
     "reflect-metadata": "0.1.13",
-    "v8-compile-cache": "2.1.1",
-    "webtorrent-health": "1.1.2",
-    "webtorrent-hybrid": "4.0.1"
+    "v8-compile-cache": "2.3.0",
+    "vlc-command": "^1.2.0",
+    "webtorrent-health": "1.2.0"
   }
 }
